+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/room/outbox/CombinedFileTransferProgress.kt
@@ -1,5 +1,6 @@
 package de.connect2x.trixnity.client.room.outbox
 
+import de.connect2x.trixnity.clientserverapi.model.media.FileTransferProgress
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.FlowCollector
@@ -7,8 +8,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.update
-import de.connect2x.trixnity.clientserverapi.model.media.FileTransferProgress
-import kotlin.collections.plus
 
 class CombinedFileTransferProgress : Flow<FileTransferProgress?> {
     private val allProgress = MutableStateFlow<Set<MutableStateFlow<FileTransferProgress?>>>(setOf())