+++ trixnity-client/src/commonTest/kotlin/de/connect2x/trixnity/client/room/RoomServiceTest.kt
@@ -9,6 +9,8 @@ import de.connect2x.trixnity.client.getInMemoryRoomStateStore
 import de.connect2x.trixnity.client.getInMemoryRoomStore
 import de.connect2x.trixnity.client.getInMemoryRoomTimelineStore
 import de.connect2x.trixnity.client.getInMemoryStickyEventStore
+import de.connect2x.trixnity.client.media.mappings.EventContentMediaMappings
+import de.connect2x.trixnity.client.media.mappings.default
 import de.connect2x.trixnity.client.mockMatrixClientServerApiClient
 import de.connect2x.trixnity.client.mocks.MediaServiceMock
 import de.connect2x.trixnity.client.mocks.RoomEventEncryptionServiceMock
@@ -99,7 +101,8 @@ class RoomServiceTest : TrixnityBaseTest() {
         matrixClientConfig = MatrixClientConfiguration(),
         typingEventHandler = TypingEventHandlerImpl(api),
         currentSyncState = CurrentSyncState(currentSyncState),
-        scope = testScope.backgroundScope
+        scope = testScope.backgroundScope,
+        eventContentMediaMappings = EventContentMediaMappings.default
     )
 
     private val eventId = EventId("\$event1")